packages/google-adk/pyproject.toml
[project]
name = "handlebar-google-adk"
version = "0.1.0"
description = "Handlebar AI governance for Google-ADK agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.13"
authors = [{ name = "Handlebar Team", email = "contact@gethandlebar.com" }]
keywords = ["governance", "ai", "agent", "llm", "guardrails", "handlebar", "safety"]
dependencies = [
    "google-adk>=1.19.0",
    "handlebar-core ==0.1.0",
]

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "litellm>=1.80.7",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
]
